[no-verify] Complete the "dual recursion context" design; still has infinite overflow.
